Add files via upload
I don't want to add the excel sheet it was writing to because someone did work hard to create it and I don't have a source for them. I made the GUI entirely myself and mostly tailored other scripts to do something similar to their original version. I pride myself in the documentation of it mostly-if anyone even sees this!